requirements.txt
streamlit==1.51.0
st-pages==1.0.1
pandas==2.2.3
plotly==5.24.1
altair==5.3.0
requests==2.32.3
urllib3==2.2.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
toml==0.10.2
python-dateutil==2.9.0.post0
httpx==0.25.2
cachetools==5.3.3
python-dotenv==1.0.1